improved code, and seperated into libraries
