+++ firmware/src/SensorLogic.h
@@ -1,0 +1,14 @@
+
+
+#pragma once
+
+#include <Arduino.h>
+#include "Config.h"
+
+int getLightValue();
+bool detectBeam();
+char detectMorseSignal();
+boolean detectPattern();
+boolean checkCount();
+
+
