--- firmware/src/Util.h
@@ -1,4 +1,0 @@
-#pragma once
-
-//Uses non blocking delay code to wait
-void wait(int miliseconds);
