+++ firmware/src/Webserver.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <ESPAsyncWebServer.h>
+#include <AsyncTCP.h>
+
+extern AsyncWebServer server;
+void setupWebServer();
+
+
